+++ tasty-dejafu/Test/Tasty/DejaFu.hs
@@ -255,7 +255,7 @@ testDejafuWay = testDejafuDiscard (const Nothing)
 
 -- | Variant of 'testDejafuWay' which can selectively discard results.
 --
--- @since unreleased
+-- @since 0.7.0.0
 testDejafuDiscard :: Show a
   => (Either Failure a -> Maybe Discard)
   -- ^ Selectively discard results.
@@ -317,7 +317,7 @@ testDejafuWayIO = testDejafuDiscardIO (const Nothing)
 
 -- | Variant of 'testDejafuDiscard' for computations which do 'IO'.
 --
--- @since unreleased
+-- @since 0.7.0.0
 testDejafuDiscardIO :: Show a => (Either Failure a -> Maybe Discard) -> Way -> MemType -> Conc.ConcIO a -> String -> Predicate a -> TestTree
 testDejafuDiscardIO discard way memtype concio name test =
   testio discard way memtype concio [(name, test)]